Fix a bug where disabling echocancel would could make it hard to disable loopback
Under: mic => echocancel => loopback

If we disable echocancel, loopback will automatically attach itself to the mic,
however the mic doesn't have the "loopback.module_id" property so the widget is
unable to tell loopback is still enabled.

To fix this, we'll disable the loopback module at the same time we disable the echo cancel module.