Rewrite ContextMenu lib to support submenu creation.
We can't use clearMenuItems() with a submenu as it will
cause a segfault when trying to remove the

ContextMenu > MenuItem > ContextMenu

It triggers the sefault when trying to remove an Action that doesn't
exist.

We fix this by creating our own removeAllItems() function which calls
removeMenuItem(item) first before item.destroy().